Added a server, basic phaser and some more file-structure
The server will contain a requestHandler that both sees to CRUD requests as well as websocket messages/upgrades. The game.js file will (for now) contain everything that is to be included in the game from sprites to movement and attacks. This will later be moved to individual files such as player.js etc